move functions into separate module
